Fixed ioctl sign-extension warnings

+++ OpenStopMotion/trunk/src/device_v4l2.cpp
@@ -38,7 +38,7 @@
 
 #define CLEAR(x) memset(&(x), 0, sizeof(x))
 
-static int xioctl(int fh, int request, void *arg)
+static int xioctl(int fh, unsigned long request, void *arg)
 {
         int r;
